Update project images and media gallery assets
Added new JPEG images to the projects assets folder and updated imports in MediaGallery and Projects components to use these new images. Adjusted the MediaGallery layout to include an additional image and modified grid styling. Also fixed the 'מי אני' menu item link in Header to point to '#intro' instead of '#about'.